+++ src/client/App.js
@@ -5,7 +5,7 @@ import { Home } from './containers/Home/Home';
 import SignIn from './containers/SignIn';
 import SignUp from './containers/SignUp';
 import ResetPassword from './containers/ResetPassword';
-import AuthenticatedRoute from './components/Auth/AuthenticatedRoute';
+import AuthenticatedRoute from './components/AuthenticatedRoute/AuthenticatedRoute';
 import { useAuthentication } from './hooks/useAuthentication';
 import Header from './components/Navigation/Header';
 import Profile from './containers/Profile';
