+++ src/client/components/AuthenticatedRoute/AuthenticatedRoute.js
