+++ src/client/containers/SignUp/index.js
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { signUp } from '../../firebase/auth';
-import UserCreationSuccess from '../../components/Success/UserCreationSuccess';
-import SignUp from '../../components/Forms/SignUp';
-import Loader from '../../components/Loader';
+import UserCreationSuccess from '../../components/UserCreationSuccess/UserCreationSuccess';
+import SignUp from '../../components/SignUp/SignUp';
+import Loader from '../../components/Loader/Loader';
 
 const getDoesPasswordsMatch = ({ password, passwordConfirm }) =>
   password === passwordConfirm;
